ObjC bindings: 6173→0 compilation errors, regen with kextract fix batch
- Regenerate all 14 framework ObjC bindings with fixed kextract
- All 6173 compilation errors resolved in kextract (see kextract commit)
- Split AppKitFunctions.kt into 15 files to avoid JVM <clinit> 64KB limit
- Increase Gradle heap to 4g for Kotlin compiler

+++ ffi/objc/src/jvmMain/kotlin/org/graphiks/kffi/objc/protocols/NSAccessibilityRow.kt
@@ -9,10 +9,10 @@ import java.lang.foreign.MemoryLayout.PathElement.*
  * Inherits protocols: NSAccessibilityGroup
  */
 interface NSAccessibilityRow : NSAccessibilityGroup {
-    fun accessibilityIndex(): NSInteger
+    fun accessibilityIndex(): Long
     
     // @optional
-    fun accessibilityDisclosureLevel(): NSInteger =
+    fun accessibilityDisclosureLevel(): Long =
         throw UnsupportedOperationException("Optional ObjC method 'accessibilityDisclosureLevel' not implemented")
     
 }
