+++ ffi/objc/src/jvmMain/kotlin/org/graphiks/kffi/objc/protocols/NSFastEnumeration.kt
@@ -8,7 +8,7 @@ import java.lang.foreign.MemoryLayout.PathElement.*
  * Kotlin/JVM interface for Objective-C protocol: NSFastEnumeration
  */
 interface NSFastEnumeration {
-    fun countByEnumeratingWithState_objects_count(state: MemorySegment, buffer: MemorySegment, len: NSUInteger): NSUInteger
+    fun countByEnumeratingWithState_objects_count(state: MemorySegment, buffer: MemorySegment, len: Long): Long
     
 }
